feat: interview result

+++ src/main/java/com/luke_industries/rest_app/models/Operation.java
@@ -11,6 +11,6 @@ import lombok.Setter;
 @Setter
 public class Operation {
     private OperationType operation;
-    private int unitCost;
+    private double unitCost;
     private int quantity;
 }
